TaskGenerator в WeatherService
1. ITaskGenerator.java удалён
2. Logger.java исправлено название переменной с `prinOutQueue` на `printOutQueue`.
3. DataSourceImple\TaskGenerator.java -> WeatherService\TaskGenerator.java убрано имплементирование от `ITaskGenerator`
4. ThreadGenerator.java изменены ссылки на TaskGenerator
5. TestServiceWeather.java изменены ссылки на TaskGenerator.
